global: split project into sub-modules
