+++ Pipfile
@@ -29,6 +29,7 @@ pytest-mock = "*"
 pytest-random-order = "*"
 pytest = "*"
 pytest-runner = "*"
+"flake8" = "*"
 
 [requires]
 python_version = "3.6"
