+++ asclepias_broker/metadata/__init__.py
@@ -4,11 +4,4 @@
 #
 # Asclepias Broker is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
-"""Asclepias broker programmatic APIs."""
-
-from .relationships import RelationshipAPI
-from .events import EventAPI
-
-__all__ = (
-    'RelationshipAPI', 'EventAPI',
-)
+"""Metadata module."""
