+++ tests/api/test_events.py
@@ -39,14 +39,14 @@ def auth_headers(access_token):
 
 def test_endpoint_auth(client):
     """Load the example events from asclepias_broker/examples."""
-    event_url = url_for('asclepias_api.event', _external=True)
+    event_url = url_for('asclepias_events.event', _external=True)
     resp = client.post(event_url, content_type='application/json')
     assert resp.status_code == 401
 
 
 def test_example_events(client, example_events, db, es_clear, auth_headers):
     """Load the example events from asclepias_broker/examples."""
-    event_url = url_for('asclepias_api.event', _external=True)
+    event_url = url_for('asclepias_events.event', _external=True)
     for data in example_events:
         resp = client.post(
             event_url, data=json.dumps(data), headers=auth_headers)
@@ -56,7 +56,7 @@ def test_example_events(client, example_events, db, es_clear, auth_headers):
 
 def test_invalid_payload(client, db, es, auth_headers):
     """Test error handling for ingestion."""
-    event_url = url_for('asclepias_api.event', _external=True)
+    event_url = url_for('asclepias_events.event', _external=True)
     # Completely invalid JSON structure
     data = {'invalid': 'true'}
     resp = client.post(event_url, data=json.dumps(data), headers=auth_headers)
