+++ tests/model/test_identifiers.py
@@ -9,8 +9,8 @@
 import pytest
 from helpers import generate_payload
 
-from asclepias_broker.api import EventAPI
-from asclepias_broker.models import Identifier
+from asclepias_broker.core.models import Identifier
+from asclepias_broker.events.api import EventAPI
 
 
 @pytest.mark.parametrize(
