+++ tests/model/test_metadata.py
@@ -10,8 +10,10 @@ import pytest
 from invenio_db import db
 from jsonschema.exceptions import ValidationError
 
-from asclepias_broker.models import Group, GroupMetadata, GroupRelationship, \
-    GroupRelationshipMetadata, GroupType, Relation
+from asclepias_broker.core.models import Relation
+from asclepias_broker.graph.models import Group, GroupRelationship, GroupType
+from asclepias_broker.metadata.models import GroupMetadata, \
+    GroupRelationshipMetadata
 
 
 def update_and_compare(m, payload, expected=None):
