+++ tests/query/test_citations.py
@@ -10,8 +10,9 @@
 import pytest
 from helpers import generate_payload
 
-from asclepias_broker.api import EventAPI, RelationshipAPI
-from asclepias_broker.models import Identifier
+from asclepias_broker.core.models import Identifier
+from asclepias_broker.events.api import EventAPI
+from asclepias_broker.search.api import RelationshipAPI
 
 TEST_CASES = [
     (
