+++ tests/schemas/test_loaders.py
@@ -7,13 +7,10 @@
 
 """Test marshmallow loaders."""
 
-from datetime import datetime
-
-import arrow
 import pytest
 from helpers import gen_identifier, gen_relation
 
-from asclepias_broker.models import Event, Identifier, Relation, Relationship
+from asclepias_broker.core.models import Identifier, Relation, Relationship
 from asclepias_broker.schemas.loaders import IdentifierSchema, \
     RelationshipSchema
